UI: Move asterisk before string to indicate "modified"
An asterisk in a UI string usually means that the attached data has
been modified, but not saved. Following the Human Interface Guidelines.

In commit 2439c6f4bb, a button was changed in the preferences to move
the asterisk from end to beginning so that it is displayed even when
the button is clipped.

This commit moves the asterisk to the beginning in four more places:
- In the brush selection menu from various painting and sculpting
  modes.
- In the Image menu from the Image editor.
- In the catalog tree view.
- In the translation add-on, when listing add-ons for which to update
  translations.

Suggested by Pablo Vazquez in !147938.

Pull Request: https://projects.blender.org/blender/blender/pulls/148565