Fix: VSE: Report text when creating new sequencer scene
The text that was shown in the report at the bottom of the
window was not formatted like other reports.
The ID name included the ID type string and wasn't in single
quotes, like is usually the case.
Also uses `TIP_` now so that this can be translated.

Pull Request: https://projects.blender.org/blender/blender/pulls/148703