Vulkan: Graphics Pipeline Compilation Improvements
This improves the graphics pipeline compilation. The previous implementation
did a lock when compiling the graphics pipelines which removed the ability
to compile multiple pipelines at the same time.

This change only locks when the pipeline cache is being read or altered.
It also has the ability to wait for another thread that is compiling the exact
same pipeline.

Pull Request: https://projects.blender.org/blender/blender/pulls/148843