Fix: Update "BBone" to "B-Bone" for consistency
The UI refers to Bendy Bones as "**B-Bones**"
in most user facing text.  (Eg - "**B-Bone** Segments")

However, it appears as "**BBone**" in the text
for the B-Bone Keying Set and its description.

This PR updates the language so it is consistent with the rest of UI.

Pull Request: https://projects.blender.org/blender/blender/pulls/149066