Fix #147876: Shortcut for Showing Controllers does not work
The Ctrl + ` shortcut does not work in pose mode does not work
because it conflicted with armature.collection_show_all which had the same hotkey.

I removed the call to armature.collection_show_all because that makes the viewport
consistent across all modes.

this was discussed in an Animation & Rigging module meeting [1]
and there were no strong feelings about the hotkey so we decided to remove it.

[1] https://devtalk.blender.org/t/2025-10-30-animation-rigging-module-meeting/43146#p-157725-patches-review-decision-time-6

Pull Request: https://projects.blender.org/blender/blender/pulls/148572