Fix #126924: Fire and smoke vorticity depends on time scale an step
Fix #126924: Fire and smoke vorticity depends on time scale an step

This makes it difficult to control these quantities separately. It causes
problems when the time scale is small, there are many time steps, or time
scale is big for static fire (without any turbulence).

This change is not backwards compatible, when the time scale and step
are different than the default. But 5.0 seems like a good time to fix
this bug.

Fix #126924
Fix #96208

Pull Request: https://projects.blender.org/blender/blender/pulls/141053