Fix #147359: use proper left alignment for output sockets with editable labels
Ideally, we could use right alignment and still be able to rename these socket
items, but that's not supported yet and is likely not trivial to implement.
Instead, we just use left-alignment.

Before this patch, the socket use a confusing mix of left and right that wasn't
useful at all.

Pull Request: https://projects.blender.org/blender/blender/pulls/149298