Fix #148788: Crash snapping with the measure tool & linked duplicates
Storing edit-mesh cache for multiple objects sharing an edit-mesh
caused a logical error where the evaluated data from each object
caused the snap cache to be considered "changed".

That change caused a use after free error (details in code-comments).

Resolve by only using the edit-mode cache for objects in edit-mode.

From a user perspective this is consistent behavior as hovering over
non-edit-meshes with the ruler active doesn't show snapping previews.

It was strange that this harped to show the preview if the other objects
harped to be linked duplicates.

Thanks to Tariq Sulley for finding the root cause of the issue.
His PR !148904 is a larger change. Prefer a smaller change for 5.0.

Ref !149321.