Fix #149154: Shader Nodes: treat unlinked inputs of muted nodes as dangling
This restores behavior from 4.5 in shader nodes where the inputs of muted nodes
are also treated as dangling. This is currently inconsistent with Geometry Nodes
but it was not an intentional change in 32c301e3cf4172b367. We can make the node
systems more consistent one way or another separately in a later release.

Pull Request: https://projects.blender.org/blender/blender/pulls/149334