EEVEE: Add back lookdev view rotation support
This is a 1:1 functionality port of the feature as it was in 4.1 and before.

The added bonus is that this now doesn't require updating the world
probe over and over. We store a copy of the world after an update and
just rotate the extracted data for each redraw. This is only done if the
rotation changes.

This option was chosen over always updating the world probe for speed.
The other solution would have been to rotate all world lighting component
during shading. But this wouldn't have worked with local probes and would
have increased the complexity of the implementation and maintenance
cost quite a lot.

So overall, the performance should be optimal and the lighting will not be
flickering like it used to in 4.1.

Caveat:
- Viewspace lighting will reupload all Volume Light Probes data each
  redraw for correctness. Best is to not have them turned on.
- Sphere lightprobes are not updated.

Pull Request: https://projects.blender.org/blender/blender/pulls/134239