Fix: Nodes: crash when adding Python-defined node
The issue was that `node_set_typeinfo` may call into Python code for custom
nodes. This in turn may trigger the node-tree-update code to run. However, this
code was missing a topology-cache-update, because the
`BKE_ntree_update_tag_node_new` was not called yet.

The fix is to simply tag the update a bit earlier.

Pull Request: https://projects.blender.org/blender/blender/pulls/149392