UI: Disable "Make Local" button for materials on linked object data
When a linked or packed material is assigned to an also linked or packed
object data data-block, the "Make Local" button (showing the linked or
packed icon) was displayed but wouldn't do anything. That's because the
object data isn't editable in this case.

This would happen commonly after making packed objects local, and trying
to make the material local to do edits.

Gray out the button and use the tooltip to explain why it's disabled,
and how to address it.

We might want to automatically change the assignment to the object in
future when pressing the button, or show a popup to let users make the
choice. But this needs further design discussion and evaluation. For now
just try to avoid confusion.

Pull Request: https://projects.blender.org/blender/blender/pulls/149507