Refactor: Vulkan: Graphics Pipeline Info
This refactor implements the structural changes to VKGraphicsInfo, that
were implemented in !148248.

The changes to `VKGraphicsInfo` provides information to design an API
 for pipeline warm-up. (part of #148398).

The VKGraphicsInfo is spliced into 3 libraries; vertex input, shaders
and fragment output. Shaders include all shaders to ensure that the
interfaces between the shaders will be compatible.

Some highlights of the changes to VKGraphicsInfo:
- Merged the pre-rasterization and fragment stage.
- Some values like vk_topology are duplicated to keep the data for
  cache-keys local.
- Building the `VkGraphicsPipelineCreateInfo` has been moved to
  `vk_graphics_pipeline.hh` as the code occupied a large part of
  `vk_pipeline_pool.cc`.

Pull Request: https://projects.blender.org/blender/blender/pulls/149509