NDOF: implement "Drone" mode
Introduce a new navigation mode, which is a slight modification to the
"Fly" mode. In the "Drone" mode, the camera will not change it's
altitude on pushing the NDOF cap forward/backward.

The altitude will change only according to up/down push.
In addition, a view leveling algorithm has been implemented,
which is useful when the user switches on the "Lock Horizon" option.
If the view is tilted, it will be leveled to match the horizon.

Ref !149058