Cleanup: replace GSet with Set in UV Select Overlap operator
The goal here is mainly to reduce the number of places where untyped GSet is used.

In theory the hash and equality function could also be added to `BVHTreeOverlap`
directly, but it's not obvious to me that the indices should always be treated as unordered.

Pull Request: https://projects.blender.org/blender/blender/pulls/149579