Vulkan: API for disabling extensions on driver level
This change adds the posibility to disable an extension on driver level
rather than enabling it, but not using it. To do so when creating the
device in `GHOST_InstanceVK::create_device` the device.extensions can
disable specific extensions based on the devices' driverId.

This will be used by !148248 to disable graphics pipeline libraries on
AMD drivers until the root of the issues have been found/addressed.

Pull Request: https://projects.blender.org/blender/blender/pulls/149625