Fix #144171: Bake Channels operator storing the frame range
When using the "Bake Channels" operator in the Graph Editor,
the frame range would get stored on a property after the first use.
That means when changing the scene frame range later, the operator
would not respond to that change.

That is fixed by adding an explicit boolean property to use the
scene range. The range property is still set to the scene range
on first run for convenience.

![image.png](/attachments/07414c8e-d0f2-449e-92b4-1f0576555e4f)

Pull Request: https://projects.blender.org/blender/blender/pulls/148988