Fix #149638: Proportional editing indicators not drawn in Dope Sheet
When using proportional editing in the Dope Sheet
the range indicator lines were no longer drawn.

The issue was that `transinfo_show_overlay` was missing
a check for the Dope Sheet and Graph Editor.

As a side effect of fixing this, the Overlay button in the
dope sheet now affects the visibility of that visualization.
If overlays are disabled, it is not shown.

Pull Request: https://projects.blender.org/blender/blender/pulls/149640