Refactor: Remove loopbreaker in BKE_fcurve_bezt_binarysearch_index_ex
The binary search of FCurves had a loopbreaker variable that would terminate a
potentially endless loop. However at closer inspection, the loop *cannot* be
endless. `start` and `end` are either `mid` + or - 1. And since `mid` is calculated
from the delta of the two, the search is guaranteed to conclude eventually
when `end > start`.

Pull Request: https://projects.blender.org/blender/blender/pulls/149691