Cleanup: Compositor: Rename compositing region to input region
This patch renames the method that retrieves the compositing region to
input region, because it is now only used for that prepose. The size of
the compositing space still relies on this method, but this will be
changed in a later patch.

Pull Request: https://projects.blender.org/blender/blender/pulls/149718