Fix: Dropping collection assets with packing applies instance offset
(See PR for steps to reproduce.)

Instead of being placed under the cursor, objects had an offset
applied that matches the instance offset of their collection. Usually we
subtract that offset to address this. This was getting the offset from
the wrong collection though.

Since 8e3214fe05, we make objects and collections local when using
packing. Apparently this changes the ID's session UID, which was already
passed to the drop operator properties. So the drop operator would fail
to find the just imported collection, and would silently fall back
(unintentionally) to the first collection in the file (!). So further
code would use the wrong collection to correct the object transforms.

Solve by making sure the operator properties get the new session UID.

Pull Request: https://projects.blender.org/blender/blender/pulls/149740