Compositor: Make Mask node aware of Display Window
This patch updates the Mask node to be aware of the display window of
the input. In practice, this means that masks will retain their position
and scale even when the camera region is moved and zoomed in the
viewport compositor.

Fixes #111344.

Pull Request: https://projects.blender.org/blender/blender/pulls/149776