Geometry Nodes: Change Index Switch Default
The current default selected data type is `Geometry` for the Index Switch &
Switch nodes. This data type is incompatible with fields, so when using
drag-to-search the user will see an error icon appear regardless of their
intended data usage. This PR changes the default to `Float` (the first data type
in the list) so that an error message isn't the first thing a user sees when the
node is added.

Potentially some smarter inferencing of whether the socket could output a field
or not could be done, but `Geometry` is already kind of an arbitrary choice so I
would say this is a better and simpler solution.

Pull Request: https://projects.blender.org/blender/blender/pulls/149422