Vulkan: Pipelines Reduction [2/2]
Currently it is possible that identical pipelines can be stored the
pipeline pool. The root cause is that state could be set that is part of
the pipeline key, but not used by the pipeline. This reduces the amount
of pipeline creations which is measurable performance hit when compiling
materials. The actual performance hit depends on the complexity of the
material and how often the unused state changed.

This change fixes this by clearing the pipeline key to only have data
that is relevant to the pipeline.

Pull Request: https://projects.blender.org/blender/blender/pulls/149821