Compositor: Retain Display Window during domain realization
This patch retains the display window of the domain in the base case
when no transformations exists. It is not clear how the display window
should be propagated when transformations exist yet. But this patch is
an improvement in most cases.

Pull Request: https://projects.blender.org/blender/blender/pulls/149844